+++ src/main/java/racingcar/domain/ResultView.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.util.Screen;
 
 import java.util.List;
