+++ src/main/java/racingcar/util/Console.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
