+++ src/test/java/racingcar/CarTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.Car;
+import racingcar.domain.CarName;
 
 import static org.assertj.core.api.Assertions.assertThat;
